options(excel_hwnd) to get specific Excel instance

+++ R/xl.get.excel.R
@@ -17,16 +17,30 @@ xl.get.excel = function()
     # return reference to Microsoft Excel
 {
 #     xls = COMCreate("Excel.Application")
-    xls = getCOMInstance("Excel.Application",force = FALSE,silent = TRUE)
-    if (is.null(xls) || ("COMErrorString" %in% class(xls))) {
-        xls = getCOMInstance("Excel.Application",force = TRUE,silent = TRUE)
-        xls[["Visible"]] = TRUE
-    } else {
-        if (!xls[["Visible"]]){
+    excel_CLSID = "{00020400-0000-0000-C000-000000000046}"
+    excel_hwnd = unlist(options("excel_hwnd"))
+    succ = FALSE
+    if(!is.null(excel_hwnd)){
+        xls = getCOMInstance_hWnd(excel_CLSID,excel_hwnd)
+        if (!(is.null(xls) || ("COMErrorString" %in% class(xls)))){
+            succ = TRUE
+            xls = xls[["Application"]]
+#             xls[["Visible"]] = TRUE
+        }
+        
+    }
+    if(!succ){
+        xls = getCOMInstance("Excel.Application",force = FALSE,silent = TRUE)
+        if (is.null(xls) || ("COMErrorString" %in% class(xls))) {
+            xls = getCOMInstance("Excel.Application",force = TRUE,silent = TRUE)
             xls[["Visible"]] = TRUE
-            warning("Connection with hidden Microsoft Excel instance. It may cause problems. Try to kill this instance from task manager.")
-        } 
-    }    
+        } else {
+            if (!xls[["Visible"]]){
+                xls[["Visible"]] = TRUE
+                warning("Connection with hidden Microsoft Excel instance. It may cause problems. Try to kill this instance from task manager.")
+            } 
+        }    
+    }
     if (xls[['workbooks']][['count']] == 0) xls[['workbooks']]$add()
     
     return(xls)
